+++ CHANGES.rst
@@ -6,7 +6,7 @@ Changelog
 
 **Fixed**
 
-- Fix writing methods on read when reindexing services
+- #1617 Fix writing methods on read when reindexing services
 
 
 1.3.4 (2020-08-11)